Moved TestDeleteAllIssues to github_test.go
We don't really distinguish between the v4 client in the tests;
we have many tests that use the DeleteIssue() API call now.
